changes
made a lot of small changes -- I've confused myself a bit since I am working with a lot of different file types (ie js when I was planning on building a react app).

I am going to build a small small project this afternoon using tailwind and go from there to see if that gives me better direction in how to fix this.